Fake Job Detection using NLP/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pydantic==2.5.3
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
joblib==1.3.2
nltk==3.8.1
jinja2==3.1.3
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.3
Pillow==10.2.0
pytesseract==0.3.10
